docker and web scripts seperated for web use
