+++ DockerSetup.go
@@ -8,71 +8,20 @@ import (
 )
 
 func main() {
-	fmt.Println("Starting GoSpoof WebUI and Docker Setup...")
+	fmt.Println("Starting GoSpoof Docker Setup...")
 
-	// 1. Check for node
-	if !checkCommand("node") {
-		fmt.Println("Node.js not found. Installing...")
-		installNode()
-	} else {
-		fmt.Println("Node.js is installed.")
-	}
-
-	// 2. Check for npm
-	if !checkCommand("npm") {
-		fmt.Println("npm not found. Please install it manually.")
-		return
-	} else {
-		fmt.Println("npm is installed.")
-	}
-
-	// 3. Navigate to Web/Server
-	serverDir := "Web/Server"
-	if _, err := os.Stat(serverDir); os.IsNotExist(err) {
-		log.Fatalf("Web/Server directory not found at %s", serverDir)
-	}
-
-	// 4. Check for Docker
+	// Check for Docker
 	if !checkCommand("docker") {
 		fmt.Println("Docker not found. Installing for Debian/Kali...")
 		installDocker()
 	} else {
-		fmt.Println("Docker is installed.")
+		fmt.Println("✅ Docker is already installed.")
 	}
-
-	// 5. Run npm setup
-	fmt.Println("Initializing npm and installing dependencies...")
-	runCmd("npm", []string{"init", "-y"}, serverDir)
-
-	runCmd("npm", []string{
-		"install",
-		"express",
-		"multer",
-		"ejs",
-		"express-ejs-layouts",
-		"socket.io",
-		"bcrypt",
-		"better-sqlite3",
-		"express-rate-limit",
-		"express-session",
-		"validator",
-	}, serverDir)
-
-	// 6. Set permissions for uploads dir
-	fmt.Println("Fixing upload directory permissions...")
-	runCmd("sudo", []string{"chmod", "-R", "755", serverDir + "/uploads"}, "")
-
-	fmt.Println("WebUI setup complete.")
 }
 
 func checkCommand(name string) bool {
 	_, err := exec.LookPath(name)
 	return err == nil
-}
-
-func installNode() {
-	runCmd("sudo", []string{"apt", "update"}, "")
-	runCmd("sudo", []string{"apt", "install", "-y", "nodejs", "npm"}, "")
 }
 
 func runCmd(cmd string, args []string, dir string) {
@@ -90,11 +39,10 @@ func runCmd(cmd string, args []string, dir string) {
 func installDocker() {
 	fmt.Println("[*] Installing Docker CLI fallback...")
 
-	// Try to install docker-cli first (lightweight fallback)
+	// Try docker-cli (lightweight option)
 	runCmd("sudo", []string{"apt", "update"}, "")
 	runCmd("sudo", []string{"apt", "install", "-y", "docker-cli", "docker-buildx"}, "")
 
-	// Check if that was enough
 	if checkCommand("docker") {
 		fmt.Println("[+] docker-cli installed and working.")
 		return
@@ -102,7 +50,7 @@ func installDocker() {
 
 	fmt.Println("[!] docker-cli not sufficient. Installing full Docker engine...")
 
-	// Full Docker Engine for Debian/Kali
+	// Full Docker Engine
 	runCmd("sudo", []string{"apt", "install", "-y", "ca-certificates", "curl", "gnupg", "lsb-release"}, "")
 	runCmd("sudo", []string{"install", "-m", "0755", "-d", "/etc/apt/keyrings"}, "")
 	runCmd("bash", []string{
@@ -116,8 +64,9 @@ func installDocker() {
 	runCmd("sudo", []string{"apt", "update"}, "")
 	runCmd("sudo", []string{"apt", "install", "-y", "docker-ce", "docker-ce-cli", "containerd.io", "docker-compose-plugin"}, "")
 
-	// Enable docker
+	// Enable Docker and add user to group
 	runCmd("sudo", []string{"systemctl", "enable", "--now", "docker"}, "")
 	runCmd("sudo", []string{"usermod", "-aG", "docker", os.Getenv("USER")}, "")
-	fmt.Println("[+] Full Docker engine installed.")
+
+	fmt.Println("[+] Full Docker engine installed and configured.")
 }
