+++ cmd/webui.go
@@ -1,0 +1,88 @@
+package main
+
+import (
+	"fmt"
+	"log"
+	"os"
+	"os/exec"
+)
+
+func main() {
+	fmt.Println("Starting GoSpoof WebUI pt.2")
+
+	// 1. Check for node
+	if !checkCommand("node") {
+		fmt.Println("Node.js not found. Installing...")
+		installNode()
+	} else {
+		fmt.Println("Node.js is installed.")
+	}
+
+	// 2. Check for npm
+	if !checkCommand("npm") {
+		fmt.Println("npm not found. Please install it manually.")
+		return
+	} else {
+		fmt.Println("npm is installed.")
+	}
+
+	// 3. Navigate to Web/Server
+	serverDir := "Web/Server"
+	if _, err := os.Stat(serverDir); os.IsNotExist(err) {
+		log.Fatalf("Web/Server directory not found at %s", serverDir)
+	}
+
+	// 5. Run npm setup
+	fmt.Println("Initializing npm and installing dependencies...")
+	runCmd("npm", []string{"init", "-y"}, serverDir)
+
+	runCmd("npm", []string{
+		"install",
+		"express",
+		"multer",
+		"ejs",
+		"express-ejs-layouts",
+		"socket.io",
+		"bcrypt",
+		"better-sqlite3",
+		"express-rate-limit",
+		"express-session",
+		"validator",
+	}, serverDir)
+	// Create uploads directory if it doesn't exist
+	uploadsDir := serverDir + "/uploads"
+	if _, err := os.Stat(uploadsDir); os.IsNotExist(err) {
+		fmt.Println("Creating uploads directory...")
+		if err := os.MkdirAll(uploadsDir, 0755); err != nil {
+			log.Fatalf("Failed to create uploads directory: %v", err)
+		}
+	}
+
+	// 6. Set permissions for uploads dir
+	fmt.Println("Fixing upload directory permissions...")
+	runCmd("sudo", []string{"chmod", "-R", "755", serverDir + "/uploads"}, "")
+
+	fmt.Println("WebUI setup complete.")
+}
+
+func checkCommand(name string) bool {
+	_, err := exec.LookPath(name)
+	return err == nil
+}
+
+func installNode() {
+	runCmd("sudo", []string{"apt", "update"}, "")
+	runCmd("sudo", []string{"apt", "install", "-y", "nodejs", "npm"}, "")
+}
+
+func runCmd(cmd string, args []string, dir string) {
+	c := exec.Command(cmd, args...)
+	if dir != "" {
+		c.Dir = dir
+	}
+	c.Stdout = os.Stdout
+	c.Stderr = os.Stderr
+	if err := c.Run(); err != nil {
+		log.Fatalf("Failed running %s: %v", cmd, err)
+	}
+}
